[fix, improve] dialog component
