+++ src/dialog/DialogTitle.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Text, StyleSheet, TextStyle, StyleProp, Platform } from 'react-native';
 import { withTheme } from '../config';
+import { RneFunctionComponent } from '../helpers';
 import { TextProps } from '../text/Text';
 
 export type DialogTitleProps = {
@@ -9,7 +10,7 @@ export type DialogTitleProps = {
   titleProps?: TextProps;
 };
 
-const DialogTitle: React.FunctionComponent<DialogTitleProps> = ({
+const DialogTitle: RneFunctionComponent<DialogTitleProps> = ({
   title,
   titleStyle,
   titleProps,
